Incorporate file storage location(s) into easypopulate_4_split.php
Somehow this was an oversight, but the $file_location had not been updated to support the removal of the admin directory from the $temp_dir variable.  This causes an error to be generated when attempting to split the files if the $temp_dir points to a location that doesn't exist (specifically if the admin folder is being used to store the files and the catalog directory does not have the foldername that is used on the admin side).