4.0.31c(cont'd) Remove 'cPath=' from exported v_html_uri category path.
Export of the v_html_uri category path (URI for a given category) was generated by zen_get_path which already includes the prefix: cPath= as part of the path generation.  The addition of 'cPath=' before the function was unnecessary and resulted in two occurrences of cPath= existing in the exported path.