Correctly loop through languages
Incorrectly used the variable $languages instead of $langcode at line 1908.  This was not a problem on the system tested because of additionally installed software.  Further the use of $langcode is an EP4 variable and should have been used instead of the standard $languages variable used elsewhere in ZC.  $langcode should have been used originally to stay within the normal structure of EP4 but was overlooked and did not present a problem (testing should have been with a clean install of ZC as well as this module.)