Prevent modifying custom field data on import if custom field identified but not included.
Original design of using custom fields always populated the field
with available information even if the field was not in the import.
This modification skips assigning a blank value to the custom
field(s) if the field is not in the import file.  It does not
prevent erasing data in the field if the field is present but
empty.