Update import categories delimiter handling
In an attempt to simplify the import/export process to use a single identifier for both in one location, the UTF-8 carat `\x5e` was used in a single variable, export and import were to both use that variable; however, the import function requires splitting the categories on whatever delimiter was to be used... The function chosen to consistently split the data into its UTF-8 parts was the mb_split function.  This function requires the delimiter to be in a regex format.  The use of `\x5e` provided as a variable without further modification does not work as a regex value because the \ needs to be escaped like `\\x5e`.  However, escaping the value up front before being used in an implode or concatenation as done for export, results in the actual letters appearing and not the UTF-8 formatted ^.  Therefore they have to be "processed" differently.  By using preg_quote() around the $categories_delimiter variable, the data within is first escaped before being used and therefore the single variable in easypopulate_4.php may remain as is and the use of that variable will be consistent between import and export.

This solution does mean that it may now be possible to add/move the delimiter to the configuration area and not be buried in the code...