v4.0.35 Add Admin Notifier after header.php
In ZC 1.5.5, loading of JQuery is included in the includes/header.php file.  As a result, previously loaded script may be overridden and made unusable/inactive.  In order to load/activate additional features, applicable script must be loaded after the header.php file. This was identified in two locations on the ZC forum one in relation to design and operation of another plugin and another in relation to this one specifically.