Provide a non-zero date on featured import of blank dates.
This commit will set the available and expire dates for featured product to the ZC standard of '0001-01-01' for each of these fields not imported instead of '0000-00-00'.

There remains other issues with import of featured product to address. Such as, the several conditions permitted by ZC for a product to remain active as a featured product.