Ensure that featured dates are established to standard
Forces the uploaded dates for featured product to either be '0001-01-01' if the entered value is less than or equal to this (blank, 0, 0000-00-00, etc..) or is set to the value uploaded if that value is greater than '0001-01-01'.  This does not ensure that the information is entered in the correct date sequence or is at all a date.  Import of a non-date will be identified in the SQL command, exchange of MM and DD may go unnoticed, but could also show an incorrect result after import.