Increase activation band of featured product and incorporate into featured insert
Previously in order to activate featured product, at least the expiration date had to be entered otherwise, featured items would be deactivated, also there would be a problem with dates if the available date was blank or 0 on insert with more recent PHP versions.  The span of dates has been increased such that automatically upon insert the status is set to active for the routine conditions that ZC already checks and for conditions that would be expected to set a featured product to be active.  This does not however automatically set a product as active if the available and expire dates are empty or '0001-01-01'.  That will need to be addressed by an admin switch as not everyone wants to have such an upload become automatically active. While this does change the span at which EP4 makes product active, it is also consistent with the ZC store's auto activate/disable.

The other change made in this commit was duplicating the changes made to the update section into the insert section for featured product.  Perhaps ideally in the future, much of this could be pulled out so that changes need only be made once, but perhaps others prefer this option.  It would just be easier to maintain if they weren't treated separately.