Restored import update of the user defined fields.
Thanks to a report in the forum and detailed troubleshooting this
issue was discovered and corrected.  It is now possible to
update the products_model field if the primary key is not the
products_model.  This was expected/intended to be possible; however,
apparently it was only implemented in the insert function not
the update.