Capture database id for insert before zen_record_admin_activity
When EP4 incorporated the zen_record_admin_activity function
to align with data capture requested/expected for ZC database
change tracking, the function was incorporated before the
insert_id was captured.  As a result, the database insertion
record location was "overwritten" by the logging function
and therefore caused tracked records to be incorrect during
import.  This has recently been pointed out as related to
the four issues identified in this commit.

Added the German installation instructions from the translated
version of this program.
Thanks to webchills of the german ZC forum for providing
instructions in German on how to install EP4.
These instructions were copied over from
easy-populate-4.028-beta-zencart-153-deutsch.zip