Update to V4.0.36.ZC (#31)
* Correct usage of bindVars

In three locations the use of bindVars was not correctly implemented. In each case the sql statement had a variable formatted like ':var:' but in the substitution performed by bindVars the variable was identified as only ':var' this causes an extra : to remain behind in the sql. This method of substitution has been chosen to prevent incorrect substitution from occurring because of sequencing of the bindVars functions when more than one variable begins with the same text as one that is fully defined by that beginning text. Ie. :value substituted into :value_me will cause an issue if :value_me isn't listed before :value in the bindVars sequencing.

* Additional bindVars mistype found

Corrected one more found mistype for using bindVars.

* Update working version of EP4

Modifications have been made here in the master. In order to minimize cross version issues, establishing this as an "Alpha" version so that when it is made effective on the ZC site without the alpha designator, users will receive notification of such an update/final version.

* Missing value to which to compare on import

if the function mb_strlen does not exist on the operating system then EP4 reverts to string length comparison by strlen; however, when this feature was added in, the comparison for $categories_name_max_len was not included and therefore all categories reported as being too long.

* Capture database id for insert before zen_record_admin_activity

When EP4 incorporated the zen_record_admin_activity function
to align with data capture requested/expected for ZC database
change tracking, the function was incorporated before the
insert_id was captured.  As a result, the database insertion
record location was "overwritten" by the logging function
and therefore caused tracked records to be incorrect during
import.  This has recently been pointed out as related to
the four issues identified in this commit.

Added the German installation instructions from the translated
version of this program.
Thanks to webchills of the german ZC forum for providing
instructions in German on how to install EP4.
These instructions were copied over from
easy-populate-4.028-beta-zencart-153-deutsch.zip

* Split out import files.

Removed code from within the easypopulate_4_import.php file to
accomplish a few things: reduce the amount of code that is
loaded on each cycle of processing, simplify understanding when
reviewing the import file, and continue streamlining the code.

Created new module files:
easypopulate_4_attrib_detailed_ep.php
easypopulate_4_default_these.php
easypopulate_4_featured_ep.php
easypopulate_4_import_categorymeta_ep.php
easypopulate_4_initiate.php
easypopulate_4_sba_detailed_ep.php
easypopulate_4_sba_stock_ep.php

* Use of alternate primary key

Primary key is now exported with full product export if not set
to the default of products_model; however, does not prevent the
primary key from being one of the other existing $filelayout
options (ie. exported twice).  While hasn't been identified
as a problem in the past, perhaps a "sanitizer" for export
should be used to ensure that a field is exported only once
and thus not complicate the import process.

Associated with issue #26

* Split out the Insert Queries for ZC 1.5.x to make future
 implementation of an update button process instead of remove
and reinstall for database affected records.

* Resequence code to improve readability.

* Remove unnecessary php ending tag.

* Clean up doc tag(s)

* Improve error reporting based on primary key.

* Correct import datatypes.

* Remove excess variables and manually clear memory.

* Correct typoes.

* Apply applicable ZC function (zen_db_input) instead of
the "generic" PHP function addslashes.

* Update for PHP/ZC versions.

* Improve Dropdown Understanding.

* Code formatting cleanup.

* Improve upload/import notification and handling.

* Improve Documentation.

* Use of alternate primary key

Primary key is now exported with full product export if not set
to the default of products_model; however, does not prevent the
primary key from being one of the other existing $filelayout
options (ie. exported twice).  While hasn't been identified
as a problem in the past, perhaps a "sanitizer" for export
should be used to ensure that a field is exported only once
and thus not complicate the import process.

Associated with issue #26

* Resequence code to improve readability.

* Remove duplicate code.

* Version Stamping/Update

* Update changelog.